+++ vscode/CHANGELOG.md
@@ -13,6 +13,15 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 
+[Unreleased]
+------------
+
+### Changed
+
+-  Server is started with process ID of Node.js client process in
+   `--clientProcessId` flag.
+
+
 [0.3.3] - 2026-05-18
 --------------------