+++ vscode/CHANGELOG.md
@@ -13,6 +13,15 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 
+[Unreleased]
+------------
+
+### Added
+
+-  Switch to *t32-language-server* v0.13.0.
+-  Update feature section in readme.
+
+
 [0.4.0] - 2026-05-22
 --------------------