+++ vscode/CHANGELOG.md
@@ -13,6 +13,14 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 
+[0.6.0] - 2026-06-17
+--------------------
+
+### Added
+
+-  Switch to *t32-language-server* v0.14.0.
+
+
 [0.5.0] - 2026-05-25
 --------------------